+++ lib/oklch.dart
@@ -1,4 +1,6 @@
-library oklch;
-
+export 'package:oklch/src/oklab_color.dart';
 export 'package:oklch/src/oklch_color.dart';
+
+export 'package:oklch/src/oklch_color_extensions.dart';
+
 export 'package:oklch/src/ui/oklch_color_picker_widget.dart';