+++ test/rgb_to_oklch_test.dart
@@ -7,8 +7,8 @@ void main() {
     /// Red color
     test('Converts RGB to OKLCH for red', () {
       // Test case 1
-      print('Test for RGB red color');
-      final OKLCHColor result = OKLCHColor.fromColor(const Color(0xffff0000));
+      const red = Color(0xffff0000);
+      final result = OKLCHColor.fromColor(red);
 
       // Verify with the expected value
       expect(result.lightness, closeTo(62.795536061455145, 0.001));
@@ -19,8 +19,7 @@ void main() {
     /// Green color
     test('Converts RGB to OKLCH for green', () {
       // Test case 2
-      print('Test for RGB green color');
-      final OKLCHColor result = OKLCHColor.fromColor(const Color(0xff00ff00));
+      final result = OKLCHColor.fromColor(const Color(0xff00ff00));
 
       // Verify with the expected value
       expect(result.lightness, closeTo(86.64396115356693, 0.001));
@@ -31,8 +30,7 @@ void main() {
     /// Blue color
     test('Converts RGB to OKLCH for blue', () {
       // Test case 3
-      print('Test for RGB blue color');
-      final OKLCHColor result = OKLCHColor.fromColor(const Color(0xff0000ff));
+      final result = OKLCHColor.fromColor(const Color(0xff0000ff));
 
       // Verify with the expected value
       expect(result.lightness, closeTo(45.201371838534286, 0.001));
@@ -44,26 +42,24 @@ void main() {
     /// https://oklch.com/#70,0.1,160,100
     test('Converts RGB to OKLCH for light green', () {
       // Test case 4
-      print('Test for RGB light green color');
-      final OKLCHColor result = OKLCHColor.fromColor(const Color(0xff62b289));
+      final result = OKLCHColor.fromColor(const Color(0xff62b289));
 
       // Verify with the expected value
-      //expect(result.lightness, closeTo(70, 0.001));
-      //expect(result.chroma, closeTo(0.1, 0.001));
-      //expect(result.hue, closeTo(160, 0.001));
+      expect(result.lightness, closeTo(70.08459661776736, 0.001));
+      expect(result.chroma, closeTo(0.1, 0.001));
+      expect(result.hue, closeTo(159.9277544974, 0.001));
     });
 
     /// White color
     /// https://oklch.com/#100,0,0,100
     test('Converts RGB to OKLCH for white', () {
       // Test case 5
-      print('Test for RGB white color');
-      final OKLCHColor result = OKLCHColor.fromColor(const Color(0xffffffff));
+      final result = OKLCHColor.fromColor(const Color(0xffffffff));
 
       // Verify with the expected value
-      //expect(result.lightness, closeTo(100, 0.001));
-      //expect(result.chroma, closeTo(0, 0.001));
-      //expect(result.hue, closeTo(0, 0.001));
+      expect(result.lightness, closeTo(100, 0.001));
+      expect(result.chroma, closeTo(0, 0.001));
+      expect(result.hue, closeTo(0, 0.001));
     });
   });
 }