+++ CHANGELOG.md
@@ -9,3 +9,10 @@ Updated formula
 ## 0.0.3
 
 Improved documentation, fix conversion iteration
+
+## 0.0.4
+
+FIX: floating point arithmetic error seen on White RBG to White OKLCH
+FEAT: OKLABColor
+FEAT: OKLCH color mix method 
+FEAT: add very good analysis